+++ rekdoc.py
@@ -1,8 +1,9 @@
 #!/bin/env python
 import re
 import sys
-from rekdoc.core import cli
+import src.rekdoc.core
+# from rekdoc.core import cli
 
 if __name__ == "__main__":
     sys.argv[0] = re.sub(r"(-script\.pyw|\.exe)?$", "", sys.argv[0])
-    sys.exit(cli())
+    sys.exit(src.rekdoc.core.cli())
