--- setup.py
@@ -1,26 +1,0 @@
-from setuptools import setup, find_packages
-import pathlib
-
-here = pathlib.Path(__file__).parent.resolve()
-# long_description = (here / "README.md").read_text(encoding="utf-8")
-
-setup(
-    name="rekdoc",
-    include_package_data=True,
-    # package_dir={"": "rekdoc/"},
-    packages=find_packages(include=["rekdoc"]),
-    # packages=['rekdoc'],
-    version="1.0.0",
-    description="A document generation program",
-    long_description="A document is generated by using a asserted data from system machines.",
-    long_description_content_type="text/plain",
-    author="Rek3000",
-    author_email="torek3k@example.com",
-    python_requires=">=3",
-    install_requires=["python-docx", "pillow", "click", "mysql.connector"],
-    entry_points={
-        "console_scripts": [
-            "rekdoc = rekdoc.core:cli",
-        ],
-    },
-)
