+++ src/rekdoc/__init__.py
@@ -1,0 +1,2 @@
+__version__ = "1.0"
+__author__ = "Rek"
